move calvinresponse to requests module
