+++ calvin/runtime/north/appmanager.py
@@ -18,11 +18,13 @@ import os
 from calvin.utilities.calvin_callback import CalvinCB
 from calvin.utilities import dynops
 from calvin.utilities import calvinlogger
-_log = calvinlogger.get_logger(__name__)
 from calvin.runtime.north.plugins.requirements import req_operations
 import calvin.utilities.calvinresponse as response
 from calvin.utilities import calvinuuid
 from calvin.actorstore.store import ActorStore, GlobalStore
+
+_log = calvinlogger.get_logger(__name__)
+
 
 class Application(object):
 
@@ -63,7 +65,7 @@ class Application(object):
         actors = {v: [k] for k, v in self.actors.items() if v is not None}
         # Collect all actors under top component name
         components = {}
-        l = (len(ns)+1) if ns else 0 
+        l = (len(ns)+1) if ns else 0
         for name, _id in actors.iteritems():
              if name.find(':',l)> -1:
                 # This is a component
@@ -102,7 +104,7 @@ class Application(object):
 
     def group_components(self):
         self.components = {}
-        l = (len(self.ns)+1) if self.ns else 0 
+        l = (len(self.ns)+1) if self.ns else 0
         for name in self.actors.values():
              if name.find(':',l)> -1:
                 # This is part of a component
@@ -114,7 +116,7 @@ class Application(object):
                     self.components[component] = [name]
 
     def component_name(self, name):
-        l = (len(self.ns)+1) if self.ns else 0 
+        l = (len(self.ns)+1) if self.ns else 0
         if name.find(':',l)> -1:
             return ':'.join(name.split(':')[0:(2 if self.ns else 1)])
         else:
@@ -309,9 +311,12 @@ class AppManager(object):
     def execute_requirements(self, application_id, cb):
         """ Build dynops iterator to collect all possible placements,
             then trigger migration.
-            
+
             For initial deployment (all actors on the current node)
         """
+    ### DEPLOYMENT ###
+
+    def deployment_add_requirements(self, application_id, reqs, cb):
         app = None
         try:
             app = self.applications[application_id]
@@ -366,7 +371,7 @@ class AppManager(object):
             else:
                 try:
                     _log.analyze(self._node.id, "+ REQ OP", {'op': req['op'], 'kwargs': req['kwargs']})
-                    it = req_operations[req['op']].req_op(self._node, 
+                    it = req_operations[req['op']].req_op(self._node,
                                             actor_id=actor_id,
                                             component=actor.component_members(),
                                             **req['kwargs']).set_name(req['op']+",SActor"+actor_id)
@@ -389,7 +394,7 @@ class AppManager(object):
         union_iters = []
         for union_req in state['req']['requirements']:
             try:
-                union_iters.append(req_operations[union_req['op']].req_op(self._node, 
+                union_iters.append(req_operations[union_req['op']].req_op(self._node,
                                         actor_id=state['actor_id'],
                                         component=state['component'],
                                         **union_req['kwargs']).set_name(union_req['op']+",UActor"+state['actor_id']))
@@ -454,7 +459,7 @@ class AppManager(object):
         """ Matrix of weights between actors how close they want to be
             0 = don't care
             1 = same node
-            
+
             Currently any nodes that are connected gets 0.5, and
             diagonal is 1:s
         """
@@ -509,7 +514,7 @@ class AppManager(object):
                                                                    actor_id=actor_id, cb=cb))
             else:
                 _log.analyze(self._node.id, "+ OTHER NODE", {'actor_id': actor_id, 'actor_name': actor_name})
-                self.storage.get_actor(actor_id, cb=CalvinCB(self._migrate_from_rt, app=app, 
+                self.storage.get_actor(actor_id, cb=CalvinCB(self._migrate_from_rt, app=app,
                                                                   actor_id=actor_id, req=req,
                                                                   move=move, cb=cb))
 
@@ -565,7 +570,7 @@ class Deployer(object):
     # TODO Make deployer use the Application class group_components, component_name and get_req
     def group_components(self):
         self.components = {}
-        l = (len(self.ns)+1) if self.ns else 0 
+        l = (len(self.ns)+1) if self.ns else 0
         for name in self.deployable['actors']:
              if name.find(':',l)> -1:
                 # This is part of a component
@@ -577,7 +582,7 @@ class Deployer(object):
                     self.components[component] = [name]
 
     def component_name(self, name):
-        l = (len(self.ns)+1) if self.ns else 0 
+        l = (len(self.ns)+1) if self.ns else 0
         if name.find(':',l)> -1:
             return ':'.join(name.split(':')[0:(2 if self.ns else 1)])
         else:
@@ -588,7 +593,7 @@ class Deployer(object):
         name = name.split(':', 1)[1] if self.ns else name
         return self.deploy_info['requirements'][name] if (self.deploy_info and 'requirements' in self.deploy_info
                                                             and name in self.deploy_info['requirements']) else []
-        
+
     def instantiate(self, actor_name, actor_type, argd, signature=None):
         """
         Instantiate an actor.
@@ -808,6 +813,6 @@ class Deployer(object):
                 c = (src_actor, src_port, dst_actor, dst_port)
                 self.connectid(c)
 
-        self.node.app_manager.finalize(self.app_id, migrate=True if self.deploy_info else False, 
+        self.node.app_manager.finalize(self.app_id, migrate=True if self.deploy_info else False,
                                        cb=CalvinCB(self.cb, deployer=self))
